AD-14. Shield. Create unit parameters with shield - refactoring to prepare unit shield - add shield object state base
