+++ Source/AtomDestiny/ObjectState/ShieldObjectStateBase.cpp
@@ -1,0 +1,1 @@
+﻿#include "ShieldObjectStateBase.h"
