+++ Source/AtomDestiny/ObjectState/ShieldObjectStateBase.h
@@ -1,0 +1,15 @@
+﻿#pragma once
+
+#include <AtomDestiny/Unit/UnitParameters.h>
+#include <AtomDestiny/ObjectState/Shield.h>
+
+#include "ShieldObjectStateBase.generated.h"
+
+UCLASS(Abstract)
+class UShieldObjectStateBase : public UUnitParameters
+{
+    GENERATED_BODY()
+    
+public:
+    
+};
